move CompletionDetails out of language
